structuring react context and place request api on context file
