--- src/components/Context/index.jsx
@@ -1,16 +1,0 @@
-import React, { createContext, useState } from 'react';
-
-// import { Container } from './styles';
-
-export const DataContext = createContext();
-
-export const DataProvider = (props) => {
-    const [data, setData] = useState({});
-    return (
-        <DataContext.Provider value={[data, setData]}>
-            {props.children}
-        </DataContext.Provider>
-    )
-}
-
-
