+++ src/components/Result/index.jsx
@@ -1,15 +1,22 @@
-import React, { useContext } from 'react';
-import { DataContext } from "../Context";
+import React, { useContext, useEffect, useState } from 'react';
+import { DataContext } from "../../Context";
 // import { Container } from './styles';
 
 const Result = () => {
-    const [data, setData] = useContext(DataContext);
+    const { data: [dataContext] } = useContext(DataContext);
+    const [dataResponse, setDataResponse] = useState({});
+
+    useEffect(() => {
+        setDataResponse(dataContext.data)
+        console.log(dataResponse)
+    }, [dataContext])
 
     return (
-       <div>
-           Result
-        {console.log(data)}
-       </div>
+        <div>
+            {dataResponse &&
+                <h1>{dataResponse.name}</h1>
+            }
+        </div>
     );
 }
