+++ src/index.js
@@ -1,6 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
+import { DataProvider } from "./Context";
 import App from './App';
 
-ReactDOM.render(<App />, document.getElementById('root'));
+ReactDOM.render(
+    <DataProvider>
+        <App />
+    </DataProvider>,
+    document.getElementById('root')
+);
